Windows: add basic support for `DoLoadImage`
Add some very basic support for DoLoadImage and UnloadImage for Windows. This
was previously not implemented and would result in a failure at runtime that is
hard to detect.

This implementation is extremely limited but serves as a starting point for
proper support for loading a library. Ideally, the user input would be
converted from UTF-8 to UTF-16. This requires additional heap allocations and
conversion logic. Error recovery there requires additional allocations both
from the local heap and the global heap.

This support enables the use of LLDB's Swift REPL on Windows.  It enables the use
of `process load <dll>` as well.