APINotes: add stub of APINotes Reader
This adds the skeleton for deserialization of the "compiled" APINotes
representation.  Rather than constantly performing string processing for
the YAML representation, we would like a (bitcode) serialized
representation of the associated metadata.  This adds the
deserialization portion of the handling.  Although the reader and writer
can now serialize and deserialize the content, it is not possible to
really access the full content yet.