[WIP/DNM] Support: introduce public API annotation support
This provides a model for annotating the "public" interfaces of a module
in LLVM.  This is motivated by the desire to have the ability to build
LLVM libraries in the shared mode to enable extraction of the shared
content between the tools.

Unlike Unix platforms, Windows has a more stringent requirement of
annotating the "dll-interface" when building a shared library.  All of
the public APIs should be marked by the appropriate `*_API` macro.  The
annotation indicates what APIs are public but are expected to live in
the dynamic library (DLL/dylib/so) that the macro is associated with.

The longer term desire here is to evolve this to support creating a
small set of libraries.  Because there is a cost associated with the
call across the image boundary (GOT+PLT/IAT), the intent is to minimise
the number of libraries.  The separation of the libraries should prefer
to split along the following disjoint sets:
- minimal set of interfaces required for `llvm-ar` (archiver/librarian)
- minimal set of interfaces required for `lld` (linker)
- minimal set of interfaces required for `clang` (compiler)

The reasoning behind this split set is that these tools are executed the
most, and minimising the penalty for these three sets is crucial for
keeping the penalty to dynamic linking (and LTO) low.  Keeping the
number of libraries low reduces the complexity in distribution and
should also help reduce the overall size of the toolchain distribution.