X86: introduce a new `FrameBuilder` Type
This introduces a new FrameBuilder type to try to reduce the
complexity in the PEI handling in the frame lowering. This is
currently a small added complexity for incrementally migrating
the implementation to this type which can be shared across the
prologue and epilogue emission as they both recompute a shared
set of information. It should allow for the prologue emission
to become a dispatch over the format - funclet, windows 32/64,
or Unix.

Differential Revision: https://reviews.llvm.org/D104557
Reviewed By: rnk