Swift: escape path literals for `ExpressionParser`
When emitting a path literal into the template for the expression
evaluation, ensure that we properly escape the native path spelling as
it may contain `\` on Windows which fails to lex in the Swift frontend,
resulting in failures.  This enables partial expression evaluation for
Windows (potentially sufficiently complete for basic `po` usage).