ObjectFile: introduce a COFF object file plugin
Windows uses COFF as an object file format and PE/COFF as an executable
file format.  They are subtly different and certain elements of a COFF
file may not be present in an executable.  Introduce a new plugin to add
support for the COFF object file format which is required to support
loading of modules built with `-gmodules`.  This is motivated by Swift
which serialises debugging information into a PCM which is a COFF object
file.